Attempt to nix stalled spider processes (#21)
Two changes to try to resolve #20 :
1. Forbid multiprocessing from re-using forked processes (maxtasksperchild=1)
2. Fixed a bug where a call to utils.time_remaining() wasn't passing back negative values (useful for allowing a small time buffer before calling pool.terminate() if the entire processing queue times out)

I'm not 100% certain that these changes have solved the problem, however I was able to reproduce having a bunch of stalled spider processes by changing the (hardcoded) utils.TIMEOUT_MINS from 11 minutes to 1 minute. It seems like doing this with the current master code causes the alarm line in spider.py to trigger, which leads to the current problem. Making the changes in this PR prevented the alarm from triggering and kept the zombie processes from showing up again.

There is going to be a performance hit from not allowing re-use of forked procs, but for something like this that is meant to run unmonitored in a cron job, it seems like trading some safety for performance is worth it.